// src/safestar/enums/safestar.enum.ts
export enum EVENT_TYPES {

  // socket actions
  SOCKET_TRACK = 'SOCKET_TRACK',
  SOCKET_TO_USER_EVENT = 'SOCKET_TO_USER_EVENT',
  SOCKET_JOIN_ROOM = 'SOCKET_JOIN_ROOM',
  SOCKET_LEAVE_ROOM = 'SOCKET_LEAVE_ROOM',

  // events
  NEW_MESSAGE = 'NEW_MESSAGE',
  NEW_MESSAGING = 'NEW_MESSAGING',
  NEW_MESSAGING_REQUEST = 'NEW_MESSAGING_REQUEST',
  MESSAGING_EVENTS_SUBSCRIBED = 'MESSAGING_EVENTS_SUBSCRIBED',
  MESSAGING_EVENTS_UNSUBSCRIBED = 'MESSAGING_EVENTS_UNSUBSCRIBED',

  NEW_TRACKER = 'NEW_TRACKER',
  NEW_TRACKER_REQUEST = 'NEW_TRACKER_REQUEST',
  NEW_TRACKER_REQUEST_CANCELED = 'NEW_TRACKER_REQUEST_CANCELED',
  NEW_TRACKER_REQUEST_ACCEPTED = 'NEW_TRACKER_REQUEST_ACCEPTED',
  NEW_TRACKER_REQUEST_REJECTED = 'NEW_TRACKER_REQUEST_REJECTED',
  STOP_TRACKING = 'STOP_TRACKING',
  STOP_TRACKER = 'STOP_TRACKER',
  
  NEW_CONVERSATION = 'NEW_CONVERSATION',
  NEW_CONVERSATION_MESSAGE = 'NEW_CONVERSATION_MESSAGE',
  CONVERSATION_MEMBER_REQUESTED = 'CONVERSATION_MEMBER_REQUESTED',
  CONVERSATION_MEMBER_REQUEST_CANCELED = 'CONVERSATION_MEMBER_REQUEST_CANCELED',
  CONVERSATION_MEMBER_REQUEST_ACCEPTED = 'CONVERSATION_MEMBER_REQUEST_ACCEPTED',
  CONVERSATION_MEMBER_REQUEST_REJECTED = 'CONVERSATION_MEMBER_REQUEST_REJECTED',
  CONVERSATION_MEMBER_ADDED = 'CONVERSATION_MEMBER_ADDED',
  CONVERSATION_MEMBER_REMOVED = 'CONVERSATION_MEMBER_REMOVED',
  CONVERSATION_MEMBER_LEFT = 'CONVERSATION_MEMBER_LEFT',
  CONVERSATION_EVENTS_SUBSCRIBED = 'CONVERSATION_EVENTS_SUBSCRIBED',
  CONVERSATION_EVENTS_UNSUBSCRIBED = 'CONVERSATION_EVENTS_UNSUBSCRIBED',
  CONVERSATION_MESSAGE_TYPING = 'CONVERSATION_MESSAGE_TYPING',
  CONVERSATION_MESSAGE_TYPING_STOPPED = 'CONVERSATION_MESSAGE_TYPING_STOPPED',
  CONVERSATION_UPDATED = 'CONVERSATION_UPDATED',
  CONVERSATION_DELETED = 'CONVERSATION_DELETED',

  NEW_WATCH = 'NEW_WATCH',
  NEW_WATCH_MESSAGE = 'NEW_WATCH_MESSAGE',
  NEW_WATCH_PULSE = 'NEW_WATCH_PULSE', // pulse within someone's watch
  WATCH_MEMBER_REQUESTED = 'WATCH_MEMBER_REQUESTED',
  WATCH_MEMBER_REQUEST_CANCELED = 'WATCH_MEMBER_REQUEST_CANCELED',
  WATCH_MEMBER_REQUEST_ACCEPTED = 'WATCH_MEMBER_REQUEST_ACCEPTED',
  WATCH_MEMBER_REQUEST_REJECTED = 'WATCH_MEMBER_REQUEST_REJECTED',
  WATCH_MEMBER_ADDED = 'WATCH_MEMBER_ADDED',
  WATCH_MEMBER_REMOVED = 'WATCH_MEMBER_REMOVED',
  WATCH_MEMBER_LEFT = 'WATCH_MEMBER_LEFT',
  WATCH_EVENTS_SUBSCRIBED = 'WATCH_EVENTS_SUBSCRIBED',
  WATCH_EVENTS_UNSUBSCRIBED = 'WATCH_EVENTS_UNSUBSCRIBED',
  WATCH_MESSAGE_TYPING = 'WATCH_MESSAGE_TYPING',
  WATCH_MESSAGE_TYPING_STOPPED = 'WATCH_MESSAGE_TYPING_STOPPED',
  WATCH_UPDATED = 'WATCH_UPDATED',
  WATCH_DELETED = 'WATCH_DELETED',

  NEW_PULSE = 'NEW_PULSE',
  NEW_PHOTO_PULSE = 'NEW_PHOTO_PULSE',
  NEW_CHECKPOINT = 'NEW_CHECKPOINT',
  NEW_CHECKPOINT_RESPONSE = 'NEW_CHECKPOINT_RESPONSE',
  NEW_USER_LOCATION_UPDATE = 'NEW_USER_LOCATION_UPDATE',
  
  TO_MESSAGING_ROOM = 'TO_MESSAGING_ROOM',
  JOIN_TO_MESSAGING_ROOM = 'JOIN_TO_MESSAGING_ROOM',
  LEAVE_TO_MESSAGING_ROOM = 'LEAVE_TO_MESSAGING_ROOM',
  MESSAGE_TYPING = 'MESSAGE_TYPING',
  MESSAGE_TYPING_STOPPED = 'MESSAGE_TYPING_STOPPED',
  
}